# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/companies": {
      "get": {
        "operationId": "CompaniesController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 25,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "default": "active",
              "type": "string",
              "enum": [
                "active",
                "deleted"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "legalName": {
                            "type": "string"
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "registrationNo": {
                            "nullable": true,
                            "type": "string"
                          },
                          "sectorCode": {
                            "nullable": true,
                            "type": "string"
                          },
                          "sizeBand": {
                            "nullable": true,
                            "type": "string",
                            "enum": [
                              "micro",
                              "small",
                              "medium",
                              "large"
                            ]
                          },
                          "employeeCount": {
                            "nullable": true,
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "country": {
                            "nullable": true,
                            "type": "string"
                          },
                          "region": {
                            "nullable": true,
                            "type": "string"
                          },
                          "website": {
                            "nullable": true,
                            "type": "string"
                          },
                          "description": {
                            "nullable": true,
                            "type": "string"
                          },
                          "address": {
                            "nullable": true,
                            "type": "object",
                            "properties": {
                              "line1": {
                                "nullable": true,
                                "type": "string",
                                "maxLength": 200
                              },
                              "line2": {
                                "nullable": true,
                                "type": "string",
                                "maxLength": 200
                              },
                              "city": {
                                "nullable": true,
                                "type": "string",
                                "maxLength": 120
                              },
                              "postalCode": {
                                "nullable": true,
                                "type": "string",
                                "maxLength": 32
                              },
                              "state": {
                                "nullable": true,
                                "type": "string",
                                "maxLength": 120
                              }
                            },
                            "additionalProperties": false
                          },
                          "logoFileId": {
                            "nullable": true,
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "fiscalYearStartMonth": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "currency": {
                            "type": "string"
                          },
                          "parentCompanyId": {
                            "nullable": true,
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          },
                          "deletedAt": {
                            "nullable": true,
                            "type": "string"
                          },
                          "restorableUntil": {
                            "nullable": true,
                            "description": "Set on deleted companies: last moment a restore is possible",
                            "type": "string"
                          },
                          "lastActivityAt": {
                            "nullable": true,
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "legalName",
                          "displayName",
                          "registrationNo",
                          "sectorCode",
                          "sizeBand",
                          "employeeCount",
                          "country",
                          "region",
                          "website",
                          "description",
                          "address",
                          "logoFileId",
                          "fiscalYearStartMonth",
                          "currency",
                          "parentCompanyId",
                          "createdAt",
                          "updatedAt",
                          "deletedAt",
                          "restorableUntil",
                          "lastActivityAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "nextCursor": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      },
      "post": {
        "operationId": "CompaniesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "legalName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "registrationNo": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 64
                  },
                  "sectorCode": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64,
                    "description": "Sector code from GET /reference/sectors"
                  },
                  "sizeBand": {
                    "type": "string",
                    "enum": [
                      "micro",
                      "small",
                      "medium",
                      "large"
                    ]
                  },
                  "employeeCount": {
                    "nullable": true,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 10000000
                  },
                  "country": {
                    "type": "string",
                    "pattern": "^[A-Z]{2}$",
                    "description": "ISO 3166-1 alpha-2 country code from GET /reference/countries"
                  },
                  "region": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 120
                  },
                  "website": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 300,
                    "format": "uri"
                  },
                  "description": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 2000
                  },
                  "address": {
                    "nullable": true,
                    "type": "object",
                    "properties": {
                      "line1": {
                        "nullable": true,
                        "type": "string",
                        "maxLength": 200
                      },
                      "line2": {
                        "nullable": true,
                        "type": "string",
                        "maxLength": 200
                      },
                      "city": {
                        "nullable": true,
                        "type": "string",
                        "maxLength": 120
                      },
                      "postalCode": {
                        "nullable": true,
                        "type": "string",
                        "maxLength": 32
                      },
                      "state": {
                        "nullable": true,
                        "type": "string",
                        "maxLength": 120
                      }
                    },
                    "additionalProperties": false
                  },
                  "fiscalYearStartMonth": {
                    "default": 1,
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 12
                  },
                  "currency": {
                    "type": "string",
                    "pattern": "^[A-Z]{3}$",
                    "description": "ISO 4217 currency code"
                  },
                  "parentCompanyId": {
                    "nullable": true,
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  }
                },
                "required": [
                  "legalName",
                  "sectorCode",
                  "sizeBand",
                  "country",
                  "fiscalYearStartMonth",
                  "currency"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "legalName": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "registrationNo": {
                      "nullable": true,
                      "type": "string"
                    },
                    "sectorCode": {
                      "nullable": true,
                      "type": "string"
                    },
                    "sizeBand": {
                      "nullable": true,
                      "type": "string",
                      "enum": [
                        "micro",
                        "small",
                        "medium",
                        "large"
                      ]
                    },
                    "employeeCount": {
                      "nullable": true,
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "country": {
                      "nullable": true,
                      "type": "string"
                    },
                    "region": {
                      "nullable": true,
                      "type": "string"
                    },
                    "website": {
                      "nullable": true,
                      "type": "string"
                    },
                    "description": {
                      "nullable": true,
                      "type": "string"
                    },
                    "address": {
                      "nullable": true,
                      "type": "object",
                      "properties": {
                        "line1": {
                          "nullable": true,
                          "type": "string",
                          "maxLength": 200
                        },
                        "line2": {
                          "nullable": true,
                          "type": "string",
                          "maxLength": 200
                        },
                        "city": {
                          "nullable": true,
                          "type": "string",
                          "maxLength": 120
                        },
                        "postalCode": {
                          "nullable": true,
                          "type": "string",
                          "maxLength": 32
                        },
                        "state": {
                          "nullable": true,
                          "type": "string",
                          "maxLength": 120
                        }
                      },
                      "additionalProperties": false
                    },
                    "logoFileId": {
                      "nullable": true,
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "fiscalYearStartMonth": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "currency": {
                      "type": "string"
                    },
                    "parentCompanyId": {
                      "nullable": true,
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "deletedAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "restorableUntil": {
                      "nullable": true,
                      "description": "Set on deleted companies: last moment a restore is possible",
                      "type": "string"
                    },
                    "lastActivityAt": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "legalName",
                    "displayName",
                    "registrationNo",
                    "sectorCode",
                    "sizeBand",
                    "employeeCount",
                    "country",
                    "region",
                    "website",
                    "description",
                    "address",
                    "logoFileId",
                    "fiscalYearStartMonth",
                    "currency",
                    "parentCompanyId",
                    "createdAt",
                    "updatedAt",
                    "deletedAt",
                    "restorableUntil",
                    "lastActivityAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      }
    },
...
    "/v1/companies/{id}/restore": {
      "post": {
        "operationId": "CompaniesController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "legalName": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "registrationNo": {
                      "nullable": true,
                      "type": "string"
                    },
                    "sectorCode": {
                      "nullable": true,
                      "type": "string"
                    },
                    "sizeBand": {
                      "nullable": true,
                      "type": "string",
                      "enum": [
                        "micro",
                        "small",
                        "medium",
                        "large"
                      ]
                    },
                    "employeeCount": {
                      "nullable": true,
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "country": {
                      "nullable": true,
                      "type": "string"
                    },
                    "region": {
                      "nullable": true,
                      "type": "string"
                    },
                    "website": {
                      "nullable": true,
                      "type": "string"
                    },
                    "description": {
                      "nullable": true,
                      "type": "string"
                    },
                    "address": {
                      "nullable": true,
                      "type": "object",
                      "properties": {
                        "line1": {
                          "nullable": true,
                          "type": "string",
                          "maxLength": 200
                        },
                        "line2": {
                          "nullable": true,
                          "type": "string",
                          "maxLength": 200
                        },
                        "city": {
                          "nullable": true,
                          "type": "string",
                          "maxLength": 120
                        },
                        "postalCode": {
                          "nullable": true,
                          "type": "string",
                          "maxLength": 32
                        },
                        "state": {
                          "nullable": true,
                          "type": "string",
                          "maxLength": 120
                        }
                      },
                      "additionalProperties": false
                    },
                    "logoFileId": {
                      "nullable": true,
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "fiscalYearStartMonth": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "currency": {
                      "type": "string"
                    },
                    "parentCompanyId": {
                      "nullable": true,
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "deletedAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "restorableUntil": {
                      "nullable": true,
                      "description": "Set on deleted companies: last moment a restore is possible",
                      "type": "string"
                    },
                    "lastActivityAt": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "legalName",
                    "displayName",
                    "registrationNo",
                    "sectorCode",
                    "sizeBand",
                    "employeeCount",
                    "country",
                    "region",
                    "website",
                    "description",
                    "address",
                    "logoFileId",
                    "fiscalYearStartMonth",
                    "currency",
                    "parentCompanyId",
                    "createdAt",
                    "updatedAt",
                    "deletedAt",
                    "restorableUntil",
                    "lastActivityAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      }
    },
...
  },
  "info": {
    "title": "ResiliSense API",
    "description": "ResiliSense 2.0 REST API. Specs: docs/revamp/modules. Errors are RFC 9457 problem+json.",
    "version": "2.0.0-alpha",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}